Added link to wiki
As it is easier to type in wiki mode while on mobile, all docs will be there for now on. This will not change when the actual code is written; this document will simply contain information on how to use Levels locally and a link to all other information about the actual language.